ENH: Allow numba aggregations to return non-float64 results (#53444)
* ENH: non float64 result support in numba groupby

* refactor & simplify

* fix CI

* maybe green?

* skip unsupported ops in other bench as well

* updates from code review

* remove commented code

* update whatsnew

* debug benchmarks

* Skip min/max benchmarks

+++ pandas/io/formats/info.py
@@ -14,8 +14,6 @@ from typing import (
     Sequence,
 )
 
-import numpy as np
-
 from pandas._config import get_option
 
 from pandas.io.formats import format as fmt
@@ -1099,4 +1097,4 @@ def _get_dataframe_dtype_counts(df: DataFrame) -> Mapping[str, int]:
     Create mapping between datatypes and their number of occurrences.
     """
     # groupby dtype.name to collect e.g. Categorical columns
-    return df.dtypes.value_counts().groupby(lambda x: x.name).sum().astype(np.intp)
+    return df.dtypes.value_counts().groupby(lambda x: x.name).sum()
